+++ internal/cortex/catalog_entity_custom_data.go
@@ -3,12 +3,15 @@ package cortex
 import (
 	"context"
 	"errors"
+	"fmt"
 	"github.com/dghubble/sling"
 )
 
 type CatalogEntityCustomDataClientInterface interface {
-	Get(ctx context.Context, entityTag string, key string) (*CatalogEntityCustomData, error)
-	List(ctx context.Context, entityTag string, params *CatalogEntityCustomDataListParams) ([]CatalogEntityCustomData, error)
+	Get(ctx context.Context, entityTag string, key string) (CatalogEntityCustomData, error)
+	List(ctx context.Context, entityTag string, params CatalogEntityCustomDataListParams) ([]CatalogEntityCustomData, error)
+	Upsert(ctx context.Context, entityTag string, req UpsertCatalogEntityCustomDataRequest) (CatalogEntityCustomData, error)
+	Delete(ctx context.Context, entityTag string, key string) error
 }
 
 type CatalogEntityCustomDataClient struct {
@@ -26,32 +29,49 @@ func (c *CatalogEntityCustomDataClient) Client() *sling.Sling {
  **********************************************************************************************************************/
 
 type CatalogEntityCustomData struct {
-	ID          string      `json:"id"`
-	Key         string      `json:"key"`
+	Tag         string      `json:"tag"` // tag of catalog entity
+	Key         string      `json:"key"` // key of custom data
 	Description string      `json:"description,omitempty"`
 	Source      string      `json:"source"`
 	Value       interface{} `json:"value"`
 	DateUpdated string      `json:"dateUpdated,omitempty"`
 }
 
+func (c *CatalogEntityCustomData) ID() string {
+	return c.Tag + ":" + c.Key
+}
+
+func (c *CatalogEntityCustomData) ValueAsString() (string, error) {
+	value := ""
+	if c.Value != nil {
+		err := error(nil)
+		value, err = InterfaceToString(c.Value)
+		if err != nil {
+			return "", err
+		}
+	}
+	return value, nil
+}
+
 /***********************************************************************************************************************
  * GET /api/v1/catalog/:tag/custom-data/:key
  **********************************************************************************************************************/
 
-func (c *CatalogEntityCustomDataClient) Get(ctx context.Context, entityTag string, key string) (*CatalogEntityCustomData, error) {
-	entityResponse := &CatalogEntityCustomData{}
-	apiError := &ApiError{}
-	response, err := c.Client().Get(Route("catalog_entities", entityTag+"/custom-data/"+key)).Receive(entityResponse, apiError)
+func (c *CatalogEntityCustomDataClient) Get(ctx context.Context, entityTag string, key string) (CatalogEntityCustomData, error) {
+	entity := CatalogEntityCustomData{}
+	apiError := ApiError{}
+	response, err := c.Client().Get(Route("catalog_entities", entityTag+"/custom-data/"+key)).Receive(&entity, &apiError)
 	if err != nil {
-		return entityResponse, errors.New("could not get catalog entity custom data: " + err.Error())
+		return entity, errors.New("could not get catalog entity custom data: " + err.Error())
 	}
 
-	err = c.client.handleResponseStatus(response, apiError)
+	err = c.client.handleResponseStatus(response, &apiError)
 	if err != nil {
-		return entityResponse, errors.Join(errors.New("Failed getting catalog entity custom data: "), err)
+		return entity, errors.Join(errors.New("Failed getting catalog entity custom data: "), err)
 	}
 
-	return entityResponse, nil
+	entity.Tag = entityTag
+	return entity, nil
 }
 
 /***********************************************************************************************************************
@@ -62,19 +82,93 @@ func (c *CatalogEntityCustomDataClient) Get(ctx context.Context, entityTag strin
 type CatalogEntityCustomDataListParams struct{}
 
 // List retrieves a list of scorecards based on a query.
-func (c *CatalogEntityCustomDataClient) List(ctx context.Context, entityTag string, params *CatalogEntityCustomDataListParams) ([]CatalogEntityCustomData, error) {
-	var entitiesResponse []CatalogEntityCustomData
-	apiError := &ApiError{}
+func (c *CatalogEntityCustomDataClient) List(ctx context.Context, entityTag string, params CatalogEntityCustomDataListParams) ([]CatalogEntityCustomData, error) {
+	var entities []CatalogEntityCustomData
+	apiError := ApiError{}
 
-	response, err := c.Client().Get(Route("catalog_entities", entityTag+"/custom-data")).QueryStruct(&params).Receive(entitiesResponse, apiError)
+	response, err := c.Client().Get(Route("catalog_entities", entityTag+"/custom-data")).QueryStruct(&params).Receive(entities, &apiError)
 	if err != nil {
 		return nil, errors.New("could not get catalog entity custom data: " + err.Error())
 	}
 
-	err = c.client.handleResponseStatus(response, apiError)
+	err = c.client.handleResponseStatus(response, &apiError)
 	if err != nil {
 		return nil, err
 	}
 
-	return entitiesResponse, nil
+	for entitiesIndex := range entities {
+		entities[entitiesIndex].Tag = entityTag
+	}
+	return entities, nil
+}
+
+/***********************************************************************************************************************
+ * POST /api/v1/catalog/:tag/custom-data
+ **********************************************************************************************************************/
+
+type UpsertCatalogEntityCustomDataRequest struct {
+	Key         string      `json:"key"`
+	Description string      `json:"description,omitempty"`
+	Value       interface{} `json:"value"`
+	Force       bool        `json:"force" url:"force,omitempty"`
+}
+
+// ToUpsertRequest https://docs.cortex.io/docs/api/add-custom-data-for-entity
+func (c *CatalogEntityCustomData) ToUpsertRequest() UpsertCatalogEntityCustomDataRequest {
+	return UpsertCatalogEntityCustomDataRequest{
+		Key:         c.Key,
+		Description: c.Description,
+		Value:       c.Value,
+	}
+}
+
+func (c *CatalogEntityCustomDataClient) Upsert(ctx context.Context, entityTag string, req UpsertCatalogEntityCustomDataRequest) (CatalogEntityCustomData, error) {
+	entity := CatalogEntityCustomData{}
+	apiError := ApiError{}
+
+	req.Force = true
+
+	body, err := c.Client().Post(Route("catalog_entities", entityTag+"/custom-data")).BodyJSON(&req).Receive(&entity, &apiError)
+	if err != nil {
+		return entity, fmt.Errorf("failed upserting custom data for entity: %+v", err)
+	}
+
+	err = c.client.handleResponseStatus(body, &apiError)
+	if err != nil {
+		return entity, err
+	}
+
+	entity.Tag = entityTag
+	return entity, nil
+}
+
+/***********************************************************************************************************************
+ * DELETE /api/v1/catalog/:tag/custom-data - Delete custom data for a catalog entity by key
+ **********************************************************************************************************************/
+
+type DeleteCatalogEntityCustomDataRequest struct {
+	Key   string `json:"key" url:"key"`
+	Force bool   `json:"force" url:"force,omitempty"`
+}
+type DeleteCatalogEntityCustomDataResponse struct{}
+
+func (c *CatalogEntityCustomDataClient) Delete(ctx context.Context, entityTag string, key string) error {
+	response := DeleteCatalogEntityCustomDataResponse{}
+	apiError := ApiError{}
+	params := DeleteCatalogEntityCustomDataRequest{
+		Key:   key,
+		Force: true,
+	}
+
+	body, err := c.Client().Delete(Route("catalog_entities", entityTag+"/custom-data")).QueryStruct(&params).Receive(&response, &apiError)
+	if err != nil {
+		return errors.New("could not delete custom data for catalog entity: " + err.Error())
+	}
+
+	err = c.client.handleResponseStatus(body, &apiError)
+	if err != nil {
+		return err
+	}
+
+	return nil
 }