+++ internal/cortex/utils.go
@@ -2,8 +2,10 @@ package cortex
 
 import (
 	"encoding/json"
+	"fmt"
 	"gopkg.in/yaml.v3"
 	"net/http"
+	"strings"
 )
 
 // yamlDecoder decodes http response YAML into a YAML-tagged struct value.
@@ -33,4 +35,34 @@ func MapFetch(m map[string]interface{}, key string, defaultValue any) any {
 
 func MapFetchToString(m map[string]interface{}, key string) string {
 	return MapFetch(m, key, "").(string)
+}
+
+func InterfaceToString(v interface{}) (string, error) {
+	value := ""
+	// see if the value is a map or a scalar
+	castedValue, ok := v.(*map[string]string)
+	if ok && castedValue != nil {
+		// this is a map, let's convert to JSON
+		sv, err := json.Marshal(castedValue)
+		if err != nil {
+			return "", err
+		}
+		value = string(sv)
+	} else {
+		value = fmt.Sprintf("%v", v)
+	}
+	return value, nil
+}
+
+func StringToInterface(v string) (interface{}, error) {
+	value := interface{}(nil)
+	if strings.Contains(v, "{") && strings.Contains(v, "}") { // hacky stupid way
+		err := json.Unmarshal([]byte(v), &value)
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		value = v
+	}
+	return value, nil
 }